refactor: restructure VS Code extension for multi-extension pack (Phase 1)
Move b2c-vs-extension to packages/vscode/core, rename to b2c-dx-core,
extract shared esbuild config, add B2CDXApi interface for inter-extension
communication, shared error handler, and context keys for dependent
extension gating. All existing features continue to work unchanged.
