+++ .vscode/launch.json
@@ -1,0 +1,16 @@
+{
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "name": "Launch B2C Extensions",
+      "type": "extensionHost",
+      "request": "launch",
+      "args": [
+        "--extensionDevelopmentPath=${workspaceFolder}/packages/vscode/core"
+      ],
+      "outFiles": [
+        "${workspaceFolder}/packages/vscode/core/dist/**/*.js"
+      ]
+    }
+  ]
+}
