+++ packages/vscode/core/.gitignore
