+++ packages/vscode/core/.prettierignore
