+++ packages/vscode/core/.vscode-test.mjs
