+++ packages/vscode/core/.vscode/extensions.json
