+++ packages/vscode/core/.vscode/launch.json
