+++ packages/vscode/core/.vscode/settings.json
