+++ packages/vscode/core/.vscode/tasks.json
