+++ packages/vscode/core/.vscodeignore
