+++ packages/vscode/core/.vscodeignorel.bak
