+++ packages/vscode/core/LICENSE
