+++ packages/vscode/core/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "b2c-vs-extension",
-  "displayName": "B2C DX VSCE",
-  "description": "VS Code extension for B2C Commerce developer experience (Page Designer assistant, B2C CLI integration)",
+  "name": "b2c-dx-core",
+  "displayName": "B2C DX Core",
+  "description": "Core B2C Commerce developer experience for VS Code — instance management, WebDAV, content libraries, log tailing, and scaffolding.",
   "version": "0.2.11",
-  "publisher": "Salesforce",
+  "publisher": "salesforce",
   "license": "Apache-2.0",
   "repository": "SalesforceCommerceCloud/b2c-developer-tooling",
-  "homepage": "https://github.com/SalesforceCommerceCloud/b2c-developer-tooling/tree/main/packages/b2c-vs-extension",
+  "homepage": "https://github.com/SalesforceCommerceCloud/b2c-developer-tooling/tree/main/packages/vscode/core",
   "bugs": "https://github.com/SalesforceCommerceCloud/b2c-developer-tooling/issues",
   "categories": [
     "Other"
@@ -746,11 +746,11 @@
     "@eslint/compat": "catalog:",
     "@types/mocha": "catalog:",
     "@types/node": "catalog:",
-    "@types/vscode": "^1.105.1",
-    "@vscode/test-cli": "^0.0.12",
-    "@vscode/test-electron": "^2.5.2",
-    "@vscode/vsce": "^3.7.1",
-    "esbuild": "^0.24.0",
+    "@types/vscode": "catalog:",
+    "@vscode/test-cli": "catalog:",
+    "@vscode/test-electron": "catalog:",
+    "@vscode/vsce": "catalog:",
+    "esbuild": "catalog:",
     "swagger-ui-dist": "^5.18.0",
     "eslint": "catalog:",
     "eslint-config-prettier": "catalog:",
