+++ packages/vscode/core/src/api-browser/api-browser-tree-provider.ts
