+++ packages/vscode/core/src/api-browser/swagger-webview.html
