+++ packages/vscode/core/src/content-tree/content-commands.ts
