+++ packages/vscode/core/src/content-tree/content-config.ts
