+++ packages/vscode/core/src/content-tree/content-fs-provider.ts
