+++ packages/vscode/core/src/content-tree/content-tree-provider.ts
