+++ packages/vscode/core/src/content-tree/index.ts
