+++ packages/vscode/core/src/logs/index.ts
