+++ packages/vscode/core/src/logs/logs-tail.ts
