+++ packages/vscode/core/src/plugins.ts
