+++ packages/vscode/core/src/sandbox-tree/index.ts
