+++ packages/vscode/core/src/sandbox-tree/sandbox-config.ts
