+++ packages/vscode/core/src/scaffold/index.ts
