+++ packages/vscode/core/src/scaffold/scaffold-commands.ts
