+++ packages/vscode/core/src/template/_app.pageId.tsx
