+++ packages/vscode/core/src/test/extension.test.ts
