+++ packages/vscode/core/tsconfig.json
