+++ packages/vscode/core/tsconfig.test.json
